+++ R/legend.R
@@ -80,7 +80,8 @@
 #'                  height = 50,
 #'                  orientation = 'vertical',
 #'                  title = htmltools::tags$div('Leaf',
-#'                                              style = 'font-size: 24px; text-align: center;'),
+#'                                              style = 'font-size: 24px;
+#'                                              text-align: center;'),
 #'                  position = 'topright')
 #'
 #'  # use raster images with size encodings
@@ -104,18 +105,20 @@
 #'        'Leaf',
 #'        style = 'font-size: 24px; text-align: center; margin-bottom: 5px;'),
 #'      position = 'topright', orientation = 'vertical')
-addLegendImage <- function(map,
-                           images,
-                           labels,
-                           title = '',
-                           labelStyle = 'font-size: 24px; vertical-align: middle;',
-                           orientation = c('vertical', 'horizontal'),
-                           width = 20,
-                           height = 20,
-                           group = NULL,
-                           className = 'info legend leaflet-control',
-                           ...) {
+addLegendImage <- function(
+    map,
+    images,
+    labels,
+    title = '',
+    labelStyle = 'font-size: 24px; vertical-align: middle;',
+    orientation = c('vertical', 'horizontal'),
+    width = 20,
+    height = 20,
+    group = NULL,
+    className = 'info legend leaflet-control',
+    ...) {
   stopifnot(length(images) == length(labels))
+  stopifnot( all(width >= 0) && all(height >= 0) )
   orientation <- match.arg(orientation)
   if ( orientation == 'vertical' ) {
     htmlTag <- htmltools::tags$div
@@ -135,11 +138,13 @@ addLegendImage <- function(map,
     f =
       function(img, label, htmlTag, height, width, maxWidth) {
         marginWidth <- max(0, (maxWidth - width) / 2)
+        imgStyle <- sprintf(
+          'vertical-align: %s; margin: %spx; margin-right: %spx; margin-left: %spx',
+          'middle', 5, marginWidth, marginWidth)
         if ( inherits(img, 'svgURI') ) {
           imgTag <- htmltools::tags$img(
             src = img,
-            style = sprintf('vertical-align: middle; margin: 5px; margin-right: %spx; margin-left: %spx',
-                            marginWidth, marginWidth),
+            style = imgStyle,
             height = height,
             width = width
           )
@@ -152,8 +157,7 @@ addLegendImage <- function(map,
               fileExt,
               base64enc::base64encode(img)
             ),
-            style = sprintf('vertical-align: middle; margin: 5px; margin-right: %spx; margin-left: %spx',
-                            marginWidth, marginWidth),
+            style = imgStyle,
             height = height,
             width = width
           )
@@ -163,12 +167,16 @@ addLegendImage <- function(map,
   )
   if (!is.null(title)) {
     htmlElements <-
-      append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
+      append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))),
+             after = 0)
   }
-  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements),
+                       className = className, group = group, ...)
 }
 
-#' Create an SVG tag for the symbol
+#' Create Map Symbols for 'leaflet' maps
+#'
+#'
 #'
 #'
 #' @param shape
@@ -211,6 +219,8 @@ addLegendImage <- function(map,
 #'
 #' HTML svg element
 #'
+#' @name mapSymbols
+#'
 #' @export
 #'
 makeSymbol <- function(shape, width, height = width, color, fillColor = color,
@@ -219,6 +229,9 @@ makeSymbol <- function(shape, width, height = width, color, fillColor = color,
   if ( 'stroke-width' %in% names(list(...)) ) {
     strokewidth <- list(...)[['stroke-width']]
   }
+  stopifnot(is.numeric(width) & is.numeric(height))
+  stopifnot(is.numeric(opacity) & is.numeric(fillOpacity))
+  stopifnot(!is.na(shape))
   shapeTag <- switch(
     shape,
     'rect' = htmltools::tags$rect(
@@ -261,7 +274,7 @@ makeSymbol <- function(shape, width, height = width, color, fillColor = color,
     ),
     'plus' = htmltools::tags$polygon(
       id = 'plus',
-      points = draw_plus(width = width, height = height, offset = strokewidth),
+      points = drawPlus(width = width, height = height, offset = strokewidth),
       stroke = color,
       fill = fillColor,
       'stroke-opacity' = opacity,
@@ -270,7 +283,7 @@ makeSymbol <- function(shape, width, height = width, color, fillColor = color,
     ),
     'cross' = htmltools::tags$polygon(
       id = 'cross',
-      points = draw_cross(width = width, height = height, offset = strokewidth),
+      points = drawCross(width = width, height = height, offset = strokewidth),
       stroke = color,
       fill = fillColor,
       'stroke-opacity' = opacity,
@@ -279,7 +292,8 @@ makeSymbol <- function(shape, width, height = width, color, fillColor = color,
     ),
     'diamond' = htmltools::tags$polygon(
       id = 'diamond',
-      points = draw_diamond(width = width, height = height, offset = strokewidth),
+      points = drawDiamond(width = width, height = height,
+                           offset = strokewidth),
       stroke = color,
       fill = fillColor,
       'stroke-opacity' = opacity,
@@ -288,7 +302,7 @@ makeSymbol <- function(shape, width, height = width, color, fillColor = color,
     ),
     'star' = htmltools::tags$polygon(
       id = 'star',
-      points = draw_star(width = width, height = height, offset = strokewidth),
+      points = drawStar(width = width, height = height, offset = strokewidth),
       stroke = color,
       fill = fillColor,
       'stroke-opacity' = opacity,
@@ -321,7 +335,8 @@ makeSymbol <- function(shape, width, height = width, color, fillColor = color,
     ),
     'polygon' = htmltools::tags$polygon(
       id = 'polygon',
-      points = draw_polygon(n = 5, width = width, height = height, offset = strokewidth),
+      points = drawPolygon(n = 5, width = width, height = height,
+                            offset = strokewidth),
       stroke = color,
       fill = fillColor,
       'stroke-opacity' = opacity,
@@ -347,51 +362,61 @@ makeSymbol <- function(shape, width, height = width, color, fillColor = color,
 makeLegendSymbol <- function(label, labelStyle, ...) {
   shapeTag <- makeSymbol(...)
   htmltools::tagList(
-    htmltools::tags$img(src = shapeTag, style = "vertical-align: middle; padding: 1px;"),
-    htmltools::tags$span(label, style = sprintf("vertical-align: middle; padding: 1px; %s", labelStyle))
+    htmltools::tags$img(src = shapeTag,
+                        style = "vertical-align: middle; padding: 1px;"),
+    htmltools::tags$span(label,
+                         style =
+                           sprintf("vertical-align: middle; padding: 1px; %s",
+                                         labelStyle))
   )
 }
 
-draw_plus <- function(width, height, offset = 0) {
+drawPlus <- function(width, height, offset = 0) {
   x <- width * c(rep(c(.4, 0, .4, .6, 1, .6), each = 2), .4) + offset
   y <- height * c(0, rep(c(.4, .6, 1, .6, .4, 0), each = 2)) + offset
   paste(x, y, sep = ',', collapse = ' ')
 }
 
-draw_diamond <- function(width, height, offset = 0) {
+drawDiamond <- function(width, height, offset = 0) {
   x <- width * c( .5, 0, .5, 1, .5) + offset
   y <- height * c(0, .5, 1, .5, 0) + offset
   paste(x, y, sep = ',', collapse = ' ')
 }
 
-draw_cross <- function(width, height, offset = 0) {
+drawCross <- function(width, height, offset = 0) {
   a <- sqrt(2) / 10
-  x <- width *  c(a, 0, .5 - a, 0, a, .5, 1 - a, 1, .5 + a, 1, 1 - a, .5, a) + offset
-  y <- height * c(0, a, .5, 1 - a, 1, .5 + a, 1, 1 - a, .5, a, 0, .5 - a, 0) + offset
+  x <- width *  c(a, 0, .5 - a, 0, a, .5, 1 - a, 1, .5 + a, 1, 1 - a, .5, a) +
+    offset
+  y <- height * c(0, a, .5, 1 - a, 1, .5 + a, 1, 1 - a, .5, a, 0, .5 - a, 0) +
+    offset
   paste(x, y, sep = ',', collapse = ' ')
 }
-draw_star <- function(width, height, offset = 0) {
-  x <- width * c(0.4,0.4,0.1414214,0,0.2585786,0,0,0.2585786,0,0.1414214,0.4,0.4,0.6,0.6,0.8585786,1,0.7414214,1,1,0.7414214,1,0.8585786,0.6,0.6,0.4) + offset
-  y <- height * c(0,0.2585786,0,0.1414214,0.4,0.4,0.6,0.6,0.8585786,1,0.7414214,1,1,0.7414214,1,0.8585786,0.6,0.6,0.4,0.4,0.1414214,0,0.2585786,0,0) + offset
+drawStar <- function(width, height, offset = 0) {
+  x <- width * c(0.4, 0.4, 0.1414214, 0, 0.2585786, 0, 0, 0.2585786, 0,
+                 0.1414214, 0.4, 0.4, 0.6, 0.6, 0.8585786, 1, 0.7414214, 1, 1,
+                 0.7414214, 1, 0.8585786, 0.6, 0.6, 0.4) + offset
+  y <- height * c(0, 0.2585786, 0, 0.1414214, 0.4, 0.4, 0.6, 0.6, 0.8585786, 1,
+                  0.7414214, 1, 1, 0.7414214, 1, 0.8585786, 0.6, 0.6, 0.4, 0.4,
+                  0.1414214, 0, 0.2585786, 0, 0) + offset
   paste(x, y, sep = ',', collapse = ' ')
 }
-draw_polygon <- function(n, width = 1, height = 1, offset = 0) {
+drawPolygon <- function(n, width = 1, height = 1, offset = 0) {
   stopifnot(n > 0 || !is.integer(n))
-  radians <- seq(-pi, pi, by = 2*pi/n)
+  radians <- seq(-pi, pi, by = 2 * pi / n)
   if ( n %% 2 == 0 ) {
-    x <- (cos(radians) + 1) * 1/2 * width + offset
-    y <- (sin(radians) + 1) * 1/2 * height + offset
+    x <- (cos(radians) + 1) * 1 / 2 * width + offset
+    y <- (sin(radians) + 1) * 1 / 2 * height + offset
   } else {
-    radians <- seq(-pi, pi, by = 2*pi/n)
-    x <- (sin(radians) + 1) * 1/2 * width + offset
-    y <- (cos(radians) + 1) * 1/2 * height + offset
+    radians <- seq(-pi, pi, by = 2 * pi / n)
+    x <- (sin(radians) + 1) * 1 / 2 * width + offset
+    y <- (cos(radians) + 1) * 1 / 2 * height + offset
   }
   paste(x, y, sep = ',', collapse = ' ')
 }
 
 #' @export
 #'
-#' @rdname makeSymbol
+#' @rdname mapSymbols
 makeSymbolIcons <- function(shape = c('rect',
                                       'circle',
                                       'triangle',
@@ -410,7 +435,6 @@ makeSymbolIcons <- function(shape = c('rect',
                             width,
                             height = width,
                             ...) {
-  shape <- match.arg(shape)
   symbols <- Map(
     makeSymbol,
     shape = shape,
@@ -424,15 +448,155 @@ makeSymbolIcons <- function(shape = c('rect',
     ...
   )
   leaflet::icons(
-    iconUrl = symbols,
+    iconUrl = unname(symbols),
     iconAnchorX = width / 2,
     iconAnchorY = height / 2
   )
 }
+#' @param map
+#'
+#' a map widget object created from 'leaflet'
+#'
+#' @param lng
+#'
+#' a numeric vector of longitudes, or a one-sided formula of the form \code{~x}
+#' where \code{x} is a variable in \code{data}; by default
+#' (if not explicitly provided), it will be automatically inferred from data
+#' by looking for a column named \code{lng}, \code{long}, or \code{longitude}
+#' (case-insensitively)
+#'
+#' @param lat
+#'
+#' a vector of latitudes or a formula (similar to the \code{lng} argument; the
+#' names \code{lat} and \code{latitude} are used when guessing the latitude
+#' column from \code{data})
+#'
+#' @param values
+#'
+#' the values used to generate shapes; can be omitted for a single type of
+#' shape
+#'
+#' @param shape
+#'
+#' the desired shape of the symbol
+#'
+#' @param color
+#'
+#' stroke color
+#'
+#' @param fillColor
+#'
+#' fill color
+#'
+#' @param opacity
+#'
+#' stroke opacity
+#'
+#' @param fillOpacity
+#'
+#' fill opacity
+#'
+#' @param strokeWidth
+#'
+#' stroke width in pixels
+#'
+#' @param width
+#'
+#' in pixels
+#'
+#' @param height
+#'
+#' in pixels
+#'
+#' @param data
+#'
+#' the data object from which the argument values are derived; by default, it
+#' is the \code{data} object provided to \code{leaflet()} initially, but can be
+#' overridden
+#'
+#' @param ...
+#'
+#' arguments to be passed to \link[leaflet]{addMarkers}
+#'
+#' @export
+#'
+#' @rdname mapSymbols
+addSymbols <- function(
+    map,
+    lng,
+    lat,
+    values,
+    shape = c("rect", "circle", "triangle", "plus", "cross", "diamond", "star"),
+    color,
+    fillColor = color,
+    opacity = 1,
+    fillOpacity = opacity,
+    strokeWidth = 1,
+    width = 20,
+    height = width,
+    data = leaflet::getMapData(map),
+    ...
+) {
+  if ( missing(values) ) {
+    shape <- match.arg(shape)
+  } else {
+    values <- as.factor(parseValues(values, data))
+    if ( length(levels(values)) > length(shape) ) {
+      stop('values has more factor levels than shape. Maximum levels is 7')
+    }
+    shape <- shape[values]
+  }
+  if ( inherits(color, 'formula') ) {
+    color <- parseValues(color, data)
+  }
+  if ( inherits(fillColor, 'formula') ) {
+    fillColor <- parseValues(fillColor, data)
+  }
+  iconSymbols <- makeSymbolIcons(shape = shape, color = color,
+                                 fillColor = fillColor, opacity = opacity,
+                                 fillOpacity = fillOpacity,
+                                 strokeWidth = strokeWidth, width = width,
+                                 height = width)
+  leaflet::addMarkers(map = map, lng = lng, lat = lat, icon = iconSymbols,
+                      data = data, ...)
+}
+#' @export
+#'
+#' @rdname mapSymbols
+addSymbolsSize <- function(
+    map,
+    lng,
+    lat,
+    values,
+    shape = c("rect", "circle", "triangle", "plus", "cross", "diamond", "star"),
+    color,
+    fillColor = color,
+    opacity = 1,
+    fillOpacity = opacity,
+    strokeWidth = 1,
+    baseSize = 20,
+    data = leaflet::getMapData(map),
+    ...
+) {
+  shape <- match.arg(shape)
+  values <- parseValues(values, data)
+  sizes <- sizeNumeric(values, baseSize)
+  if ( inherits(color, 'formula') ) {
+    color <- parseValues(color, data)
+  }
+  if ( inherits(fillColor, 'formula') ) {
+    fillColor <- parseValues(fillColor, data)
+  }
+  addSymbols(map = map, lng = lng, lat = lat, shape = shape, color = color,
+             fillColor = fillColor, opacity = opacity,
+             fillOpacity = fillOpacity, strokeWidth = strokeWidth,
+             width = sizes, data = data, ...)
+}
 
 #' Add Customizable Color Legends to a 'leaflet' map widget
 #'
-#' Functions for more control over the styling of 'leaflet' legends. The 'leaflet'
+#' Functions for more control over the styling of 'leaflet' legends.
+#' The 'leaflet'
 #' map is passed through and the output is a 'leaflet' control so that
 #' the legends are integrated with other functionality of the API. Style
 #' the text of the labels, the symbols used, orientation of the legend items,
@@ -452,7 +616,8 @@ makeSymbolIcons <- function(shape = c('rect',
 #'
 #' @param bins
 #'
-#' an approximate number of tick-marks on the color gradient for the colorNumeric palette
+#' an approximate number of tick-marks on the color gradient for the
+#' colorNumeric palette
 #'
 #' @param title
 #'
@@ -509,6 +674,14 @@ makeSymbolIcons <- function(shape = c('rect',
 #' @param className
 #'
 #' extra CSS class to append to the control, space separated
+#'
+#' @param data a data object. Currently supported objects are matrices, data
+#'   frames, spatial objects from the \pkg{sp} package
+#'   (\code{SpatialPoints}, \code{SpatialPointsDataFrame}, \code{Polygon},
+#'   \code{Polygons}, \code{SpatialPolygons}, \code{SpatialPolygonsDataFrame},
+#'   \code{Line}, \code{Lines}, \code{SpatialLines}, and
+#'   \code{SpatialLinesDataFrame}), and
+#'   spatial data frames from the \pkg{sf} package.
 #'
 #' @param ...
 #'
@@ -604,7 +777,8 @@ makeSymbolIcons <- function(shape = c('rect',
 #'   ) %>%
 #'   addLegendFactor(
 #'     pal = factorPal,
-#'     title = htmltools::tags$div('addLegendFactor', style = 'font-size: 24px; color: red;'),
+#'     title = htmltools::tags$div('addLegendFactor', style = 'font-size: 24px;
+#'     color: red;'),
 #'     values = quakes$group,
 #'     position = 'topright',
 #'     shape = 'triangle',
@@ -631,7 +805,6 @@ makeSymbolIcons <- function(shape = c('rect',
 #'   ) %>%
 #'   addLegendBin(
 #'     pal = binPal,
-#'     values = quakes$mag,
 #'     position = 'topright',
 #'     title = 'addLegendBin',
 #'     labelStyle = 'font-size: 18px; font-weight: bold;',
@@ -662,7 +835,6 @@ makeSymbolIcons <- function(shape = c('rect',
 #'   ) %>%
 #'   addLegendBin(
 #'     pal = binPal,
-#'     values = quakes$mag,
 #'     position = 'bottomleft',
 #'     title = 'addLegendBin',
 #'     group = 'Bin'
@@ -681,16 +853,28 @@ addLegendNumeric <- function(map,
                              width = 20,
                              height = 100,
                              bins = 7,
-                             numberFormat = function(x) {prettyNum(x, format = 'f', big.mark = ',', digits = 3, scientific = FALSE)},
+                             numberFormat = function(x) {
+                               prettyNum(x, format = 'f', big.mark = ',',
+                                         digits = 3, scientific = FALSE)
+                               },
                              tickLength = 4,
                              tickWidth = 1,
                              decreasing = FALSE,
                              fillOpacity = 1,
                              group = NULL,
                              className = 'info legend leaflet-control',
+                             data = leaflet::getMapData(map),
                              ...) {
   stopifnot( attr(pal, 'colorType') == 'numeric' )
+  stopifnot( is.numeric(width) && is.numeric(height) && width >= 0 &&
+               height >= 0 )
+  stopifnot( is.numeric(tickLength) && is.numeric(tickWidth) &&
+               tickLength >= 0 && tickWidth >= 0 )
   shape <- match.arg(shape)
+  id <- sprintf('gradient-%s-%d',
+                gsub('[[:punct:]]|\\s', '', deparse(match.call()[['values']])),
+                length(map[["x"]][["calls"]]) + 1)
+  values <- parseValues(values = values, data = data)
   rng <- range(values, na.rm = TRUE)
   breaks <- pretty(values, bins)
   if ( breaks[1] < rng[1] ) {
@@ -704,7 +888,7 @@ addLegendNumeric <- function(map,
   offsets <- sprintf('%f%%', scaledbreaks * 100)
   invisible(lapply(c('x1', 'y1', 'x2', 'y2'), assign, 0, pos = environment()))
   orientation <- match.arg(orientation)
-  vertical <- orientation =='vertical'
+  vertical <- orientation == 'vertical'
   outer <- c(1, length(breaks))
   if ( vertical ) {
     labels <- breaks[-outer]
@@ -714,22 +898,22 @@ addLegendNumeric <- function(map,
   if ( decreasing ) {
     labels <- rev(labels)
   }
-  if ( vertical & decreasing ) {
+  if ( vertical && decreasing ) {
     y1 <- 1
-  } else if ( vertical & !decreasing ) {
+  } else if ( vertical && !decreasing ) {
     y2 <- 1
-  } else if ( !vertical & decreasing ) {
+  } else if ( !vertical && decreasing ) {
     x1 <- 1
   } else {
     x2 <- 1
   }
   labels <- numberFormat(labels)
-  #fontSize <- as.numeric(sub('.*(font-size: )([0-9]+)(px).*', '\\2', labelStyle))
   labelStyle <- ''
-  cexAdj <- 1#fontSize / 14
-  textWidth <- max(graphics::strwidth(labels, units = 'inches', cex = cexAdj)) * 72
-  textHeight <- max(graphics::strheight(labels, units = 'inches', cex = cexAdj)) * 72
-  padLabel <- 5
+  cexAdj <- 1
+  textWidth <- max(graphics::strwidth(labels, units = 'inches',
+                                      cex = cexAdj)) * 72
+  textHeight <- max(graphics::strheight(labels, units = 'inches',
+                                        cex = cexAdj)) * 72
   if ( vertical ) {
     svgwidth <- width + tickLength
     svgheight <- height
@@ -738,7 +922,6 @@ addLegendNumeric <- function(map,
     linex2 <- width + tickLength
     liney1 <- scaledbreaks[-outer] * height
     liney2 <- scaledbreaks[-outer] * height
-    textx <- 0
     texty <- scaledbreaks[-outer] * height
   } else {
     svgwidth <- width
@@ -751,12 +934,11 @@ addLegendNumeric <- function(map,
   }
   if ( shape == 'rect' ) {
     rectround <- list(rx = '0%')
-  } else if ( shape == 'stadium' & vertical ) {
+  } else if ( shape == 'stadium' && vertical ) {
     rectround <- list(rx = '10%')
   } else {
     rectround <- list(ry = '10%')
   }
-  id <- paste0(sample(c(0:9, letters, LETTERS), 10, replace = TRUE), collapse = '')
   svgElement <- htmltools::tags$svg(
     width = svgwidth,
     height = svgheight,
@@ -789,37 +971,51 @@ addLegendNumeric <- function(map,
   )
   if ( vertical ) {
     htmlElements <- list(htmltools::tags$div(style = 'display: flex;',
-      htmltools::tags$div(svgElement, style="margin-right: 5px"),
-      htmltools::tags$div(style=sprintf("width: %spx; height: %spx; position: relative; %s", textWidth, height, labelStyle),
-                          class="container",
-                          Map(function(y, label, leftPos) {
-                            htmltools::tags$div(style=sprintf("position:absolute; left:%spx; top: %spx;",
-                                                              leftPos,
-                                                              y),
-                                                label)
-                          },
-                          texty - textHeight , labels, textWidth - graphics::strwidth(labels, units = 'inches', cex = cexAdj) * 72)
+      htmltools::tags$div(svgElement, style = "margin-right: 5px"),
+      htmltools::tags$div(
+        style = sprintf("width: %spx; height: %spx; position: relative; %s",
+                        textWidth, height, labelStyle),
+        class = "container",
+        Map(function(y, label, leftPos) {
+          htmltools::tags$div(
+            style = sprintf("position:absolute; left:%spx; top: %spx;",
+                            leftPos, y), label)
+          },
+          y = texty - textHeight,
+          label = labels,
+          leftPos = textWidth - graphics::strwidth(labels, units = 'inches',
+                                                   cex = cexAdj) * 72)
       )
-      ,htmltools::tags$div(style="width: 8px; position: relative;")
+      , htmltools::tags$div(style = "width: 8px; position: relative;")
     ))
   } else {
     htmlElements <- list(
-      htmltools::tags$div(style = sprintf('margin-right: %spx; margin-left: %spx', textWidth / 2, textWidth / 2 ), svgElement),
-      htmltools::tags$div(style=sprintf("width: 100%%; height: 1rem; position: relative; %s", labelStyle),
-                          htmltools::tags$div(style=sprintf("position:absolute; left:%spx; top: 0px;", 0),
-                                              labels[1]),
-                          htmltools::tags$div(style=sprintf("position:absolute; left:%spx; top: 0px;",
-                                                            width - diff(graphics::strwidth(labels, units = 'inches', cex = cexAdj)) * 72),
-                                              labels[2])
+      htmltools::tags$div(
+        style = sprintf('margin-right: %spx; margin-left: %spx', textWidth / 2,
+                        textWidth / 2 ), svgElement),
+      htmltools::tags$div(
+        style = sprintf("width: 100%%; height: 1rem; position: relative; %s",
+                        labelStyle),
+        htmltools::tags$div(
+          style = sprintf("position:absolute; left:%spx; top: 0px;", 0),
+          labels[1]),
+        htmltools::tags$div(
+          style = sprintf("position:absolute; left:%spx; top: 0px;",
+                        width - diff(graphics::strwidth(labels,
+                                                        units = 'inches',
+                                                        cex = cexAdj)) * 72),
+          labels[2])
 
       )
       )
   }
   if ( !is.null(title) ) {
     htmlElements <-
-      append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
+      append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))),
+             after = 0)
   }
-  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements),
+                       className = className, group = group, ...)
 }
 
 #' @export
@@ -831,19 +1027,28 @@ addLegendQuantile <- function(map,
                               values,
                               title = NULL,
                               labelStyle = '',
-                              shape = c('rect', 'circle', 'triangle', 'plus', 'cross', 'diamond', 'star', 'stadium', 'line', 'polygon'),
+                              shape = c('rect', 'circle', 'triangle', 'plus',
+                                        'cross', 'diamond', 'star', 'stadium',
+                                        'line', 'polygon'),
                               orientation = c('vertical', 'horizontal'),
                               width = 24,
                               height = 24,
-                              numberFormat = function(x) {prettyNum(x, big.mark = ',', scientific = FALSE, digits = 1)},
+                              numberFormat = function(x) {
+                                prettyNum(x, big.mark = ',', scientific = FALSE,
+                                          digits = 1)
+                                },
                               opacity = 1,
                               fillOpacity = opacity,
                               group = NULL,
                               className = 'info legend leaflet-control',
+                              data = leaflet::getMapData(map),
                               ...) {
   stopifnot( attr(pal, 'colorType') == 'quantile' )
+  stopifnot( width >= 0 && height >= 0 )
   shape <- match.arg(shape)
+  orientation <- match.arg(orientation)
   probs <- attr(pal, 'colorArgs')[['probs']]
+  values <- parseValues(values = values, data = data)
   if ( is.null(numberFormat) ) {
     labels <- sprintf(' %3.0f%% - %3.0f%%',
                       probs[-length(probs)] * 100,
@@ -859,6 +1064,106 @@ addLegendQuantile <- function(map,
                     labels[-1])
   }
   colors <- unique(pal(sort(values)))
+  htmlElements <- makeLegendCategorical(shape = shape, labels = labels,
+                                        colors = colors,
+                                        labelStyle = labelStyle,
+                                        height = height, width = width,
+                                        opacity = opacity,
+                                        fillOpacity = fillOpacity,
+                                        orientation = orientation,
+                                        title = title)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements),
+                       className = className, group = group, ...)
+}
+
+#' @export
+#'
+#' @rdname addLeafLegends
+#'
+addLegendBin <- function(map,
+                         pal,
+                         values,
+                         title = NULL,
+                         labelStyle = '',
+                         shape = c('rect', 'circle', 'triangle', 'plus',
+                                   'cross', 'diamond', 'star', 'stadium',
+                                   'line', 'polygon'),
+                         orientation = c('vertical', 'horizontal'),
+                         width = 24,
+                         height = 24,
+                         numberFormat = function(x) {
+                           format(round(x, 3), big.mark = ',', trim = TRUE,
+                                  scientific = FALSE)
+                         },
+                         opacity = 1,
+                         fillOpacity = opacity,
+                         group = NULL,
+                         className = 'info legend leaflet-control',
+                         data = leaflet::getMapData(map),
+                         ...) {
+  stopifnot( attr(pal, 'colorType') == 'bin' )
+  stopifnot( width >= 0 && height >= 0 )
+  shape <- match.arg(shape)
+  orientation <- match.arg(orientation)
+  bins <- attr(pal, 'colorArgs')[['bins']]
+  labels <- sprintf(' %s - %s', numberFormat(bins[-length(bins)]),
+                    numberFormat(bins[-1]))
+  colors <- pal((bins[-1] + bins[-length(bins)]) / 2 )
+  htmlElements <- makeLegendCategorical(shape = shape, labels = labels,
+                                        colors = colors,
+                                        labelStyle = labelStyle,
+                                        height = height, width = width,
+                                        opacity = opacity,
+                                        fillOpacity = fillOpacity,
+                                        orientation = orientation,
+                                        title = title)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements),
+                       className = className, group = group, ...)
+}
+
+#' @export
+#'
+#' @rdname addLeafLegends
+#'
+addLegendFactor <- function(map,
+                            pal,
+                            values,
+                            title = NULL,
+                            labelStyle = '',
+                            shape = c('rect', 'circle', 'triangle', 'plus',
+                                      'cross', 'diamond', 'star', 'stadium',
+                                      'line', 'polygon'),
+                            orientation = c('vertical', 'horizontal'),
+                            width = 24,
+                            height = 24,
+                            opacity = 1,
+                            fillOpacity = opacity,
+                            group = NULL,
+                            className = 'info legend leaflet-control',
+                            data = leaflet::getMapData(map),
+                            ...) {
+  stopifnot( attr(pal, 'colorType') == 'factor' )
+  stopifnot( width >= 0 && height >= 0 )
+  shape <- match.arg(shape)
+  orientation <- match.arg(orientation)
+  values <- parseValues(values = values, data = data)
+  values <- sort(unique(values))
+  labels <- sprintf(' %s', values)
+  colors <- pal(values)
+  htmlElements <- makeLegendCategorical(shape = shape, labels = labels,
+                                        colors = colors,
+                                        labelStyle = labelStyle,
+                                        height = height, width = width,
+                                        opacity = opacity,
+                                        fillOpacity = fillOpacity,
+                                        orientation = orientation,
+                                        title = title)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements),
+                       className = className, group = group, ...)
+}
+
+makeLegendCategorical <- function(shape, labels, colors, labelStyle, height,
+                              width, opacity, fillOpacity, orientation, title) {
   htmlElements <- Map(
     f = makeLegendSymbol,
     shape = shape,
@@ -871,105 +1176,19 @@ addLegendQuantile <- function(map,
     fillOpacity = fillOpacity,
     'stroke-width' = 1
   )
-  orientation <- match.arg(orientation)
   if ( orientation == 'vertical' ) {
-    htmlElements <- lapply(htmlElements, htmltools::tagList, htmltools::tags$br())
+    htmlElements <- lapply(htmlElements, htmltools::tagList,
+                           htmltools::tags$br())
   }
   if (!is.null(title)) {
     htmlElements <-
-      append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
+      append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))),
+             after = 0)
   }
-  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
+  htmlElements
 }
 
-#' @export
-#'
-#' @rdname addLeafLegends
-#'
-addLegendBin <- function(map,
-                         pal,
-                         values,
-                         title = NULL,
-                         labelStyle = '',
-                         shape = c('rect', 'circle', 'triangle', 'plus', 'cross', 'diamond', 'star', 'stadium', 'line', 'polygon'),
-                         orientation = c('vertical', 'horizontal'),
-                         width = 24,
-                         height = 24,
-                         opacity = 1,
-                         fillOpacity = opacity,
-                         group = NULL,
-                         className = 'info legend leaflet-control',
-                         ...) {
-  stopifnot( attr(pal, 'colorType') == 'bin' )
-  shape <- match.arg(shape)
-  bins <- prettyNum(attr(pal, 'colorArgs')[['bins']], format = 'f', big.mark = ',', scientific = FALSE)
-  labels <- sprintf(' %s - %s', bins[-length(bins)], bins[-1])
-  colors <- unique(pal(sort(values)))
-  htmlElements <- Map(f = makeLegendSymbol,
-      shape = shape,
-      label = labels,
-      color = colors,
-      labelStyle = labelStyle,
-      height = height,
-      width = width,
-      opacity = opacity,
-      fillOpacity = fillOpacity,
-      'stroke-width' = 1)
-  orientation <- match.arg(orientation)
-  if ( orientation == 'vertical' ) {
-    htmlElements <- lapply(htmlElements, htmltools::tagList, htmltools::tags$br())
-  }
-  if (!is.null(title)) {
-    htmlElements <-
-      append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
-  }
-  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
-}
-
-#' @export
-#'
-#' @rdname addLeafLegends
-#'
-addLegendFactor <- function(map,
-                            pal,
-                            values,
-                            title = NULL,
-                            labelStyle = '',
-                            shape = c('rect', 'circle', 'triangle', 'plus', 'cross', 'diamond', 'star', 'stadium', 'line', 'polygon'),
-                            orientation = c('vertical', 'horizontal'),
-                            width = 24,
-                            height = 24,
-                            opacity = 1,
-                            fillOpacity = opacity,
-                            group = NULL,
-                            className = 'info legend leaflet-control',
-                            ...) {
-  stopifnot( attr(pal, 'colorType') == 'factor' )
-  shape <- match.arg(shape)
-  labels <- sprintf(' %s', sort(unique(values)))
-  colors <- pal(sort(unique(values)))
-  htmlElements <- Map(f = makeLegendSymbol,
-      shape = shape,
-      label = labels,
-      color = colors,
-      labelStyle = labelStyle,
-      height = height,
-      width = width,
-      opacity = opacity,
-      fillOpacity = fillOpacity,
-      'stroke-width' = 1)
-  orientation <- match.arg(orientation)
-  if ( orientation == 'vertical' ) {
-    htmlElements <- lapply(htmlElements, htmltools::tagList, htmltools::tags$br())
-  }
-  if (!is.null(title)) {
-    htmlElements <-
-      append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
-  }
-  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
-}
-
-#' Add a legend that for the sizing of symbols or the width of lines
+#' Add a legend for the sizing of symbols or the width of lines
 #'
 #' @param map
 #'
@@ -1032,10 +1251,6 @@ addLegendFactor <- function(map,
 #'
 #' width of symbol outline
 #'
-#' @param colorValues
-#'
-#' the values used to generate color from the palette function
-#'
 #' @param numberFormat
 #'
 #' formatting functions for numbers that are displayed e.g. format, prettyNum
@@ -1048,6 +1263,14 @@ addLegendFactor <- function(map,
 #'
 #' extra CSS class to append to the control, space separated
 #'
+#' @param data a data object. Currently supported objects are matrices, data
+#'   frames, spatial objects from the \pkg{sp} package
+#'   (\code{SpatialPoints}, \code{SpatialPointsDataFrame}, \code{Polygon},
+#'   \code{Polygons}, \code{SpatialPolygons}, \code{SpatialPolygonsDataFrame},
+#'   \code{Line}, \code{Lines}, \code{SpatialLines}, and
+#'   \code{SpatialLinesDataFrame}), and
+#'   spatial data frames from the \pkg{sf} package.
+#'
 #' @param ...
 #'
 #' arguments to pass to
@@ -1056,7 +1279,7 @@ addLegendFactor <- function(map,
 #'
 #' \link[base]{pretty} for sizeBreaks
 #'
-#' \link[leaflegend]{makeSymbol} for makeSizeIcons
+#' \link[leaflegend]{makeSymbol} for makeSymbolsSize
 #'
 #' @return
 #'
@@ -1064,6 +1287,7 @@ addLegendFactor <- function(map,
 #'
 #' @export
 #'
+#' @name legendSymbols
 #'
 #' @examples
 #' library(leaflet)
@@ -1095,10 +1319,10 @@ addLegendFactor <- function(map,
 #'
 #' # a wrapper for making icons is provided
 #' sizeSymbols <-
-#' makeSizeIcons(
+#' makeSymbolsSize(
 #'   quakes$depth,
 #'   shape = 'cross',
-#'   pal = numPal,
+#'   fillColor = numPal(quakes$depth),
 #'   color = 'black',
 #'   strokeWidth = 1,
 #'   opacity = .8,
@@ -1161,20 +1385,27 @@ addLegendSize <- function(map,
                           values,
                           title = NULL,
                           labelStyle = '',
-                          shape = c('rect', 'circle', 'triangle', 'plus', 'cross', 'diamond', 'star', 'stadium'),
+                          shape = c('rect', 'circle', 'triangle', 'plus',
+                                    'cross', 'diamond', 'star', 'stadium',
+                                    'polygon'),
                           orientation = c('vertical', 'horizontal'),
                           color,
-                          fillColor,
+                          fillColor = color,
                           strokeWidth = 1,
                           opacity = 1,
                           fillOpacity = opacity,
                           breaks = 5,
-                          baseSize = 10,
-                          numberFormat = function(x) {prettyNum(x, big.mark = ',', scientific = FALSE, digits = 1)},
+                          baseSize = 20,
+                          numberFormat = function(x) {
+                            prettyNum(x, big.mark = ',', scientific = FALSE,
+                                      digits = 1)
+                            },
                           group = NULL,
                           className = 'info legend leaflet-control',
+                          data = leaflet::getMapData(map),
                           ...) {
-  shape <- match.arg(shape)
+  shape <- match.arg(shape) # is this necessary
+  values <- parseValues(values = values, data = data)
   sizes <- sizeBreaks(values, breaks, baseSize)
   if ( missing(color) ) {
     stopifnot( missing(color) & !missing(pal))
@@ -1202,7 +1433,8 @@ addLegendSize <- function(map,
                  opacity = opacity,
                  fillOpacity = fillOpacity,
                  `stroke-width` = strokeWidth)
-  addLegendImage(map, images = symbols, labels = numberFormat(as.numeric(names(sizes))),
+  addLegendImage(map, images = symbols,
+                 labels = numberFormat(as.numeric(names(sizes))),
                  title = title, labelStyle = labelStyle,
                  orientation = orientation, width = sizes, height = sizes,
                  group = group, className = className, ...)
@@ -1211,15 +1443,30 @@ addLegendSize <- function(map,
 
 #' @export
 #'
-#' @rdname addLegendSize
+#' @rdname mapSymbols
 sizeNumeric <- function(values, baseSize) {
+  stopifnot(baseSize > 0)
   values / mean(values, na.rm = TRUE) * baseSize
 }
 
+#' @param breaks
+#'
+#' an integer specifying the number of breaks or a numeric vector of the breaks
+#'
+#' @param baseSize
+#'
+#' re-scaling size in pixels of the mean of the values, the average value will
+#' be this exact size
+#'
+#' @param ...
+#'
+#' arguments to pass to \code{pretty}
+#'
 #' @export
 #'
-#' @rdname addLegendSize
+#' @rdname mapSymbols
 sizeBreaks <- function(values, breaks, baseSize, ...) {
+  stopifnot(baseSize > 0)
   if ( length(breaks) == 1 ) {
     breaks <- pretty(values, breaks, ...)
   }
@@ -1229,54 +1476,30 @@ sizeBreaks <- function(values, breaks, baseSize, ...) {
 
 #' @export
 #'
-#' @rdname addLegendSize
-makeSizeIcons <- function(values,
+#' @rdname mapSymbols
+makeSymbolsSize <- function(values,
                           shape = c('rect', 'circle', 'triangle', 'plus',
-                                    'cross', 'diamond', 'star', 'stadium'),
-                          pal,
+                                    'cross', 'diamond', 'star', 'stadium',
+                                    'polygon'),
                           color,
-                          colorValues,
                           fillColor,
-                          opacity,
+                          opacity = 1,
                           fillOpacity = opacity,
                           strokeWidth = 1,
                           baseSize,
                           ...
                           ) {
   shape <- match.arg(shape)
-  if ( missing(color) ) {
-    if ( missing(colorValues) ) {
-      stopifnot( all(missing(color), missing(pal), missing(colorValues)) )
-      colors <- pal(values)
-    } else {
-      colors <- pal(colorValues)
-    }
-  } else {
-    stopifnot(length(color) == 1 || length(color) == length(values))
-    colors <- color
-  }
-  if ( missing(fillColor) ) {
-    if ( !missing(pal) ) {
-      if ( missing(colorValues) ) {
-        fillColors <- pal(values)
-      } else {
-        fillColors <- pal(colorValues)
-      }
-    } else {
-      fillColors <- colors
-    }
-  } else {
-    stopifnot(length(fillColor) == 1 || length(fillColor) == length(values))
-    fillColors <- fillColor
-  }
-
+  stopifnot(strokeWidth >= 0)
+  stopifnot(length(color) == 1 || length(color) == length(values))
+  stopifnot(length(fillColor) == 1 || length(fillColor) == length(values))
   sizes <- sizeNumeric(values, baseSize)
   makeSymbolIcons(
     shape = shape,
     width = sizes,
     height = sizes,
-    color = colors,
-    fillColor = fillColors,
+    color = color,
+    fillColor = fillColor,
     opacity = opacity,
     fillOpacity = fillOpacity,
     strokeWidth = strokeWidth,
@@ -1289,7 +1512,7 @@ makeSizeIcons <- function(values,
 #'
 #' @export
 #'
-#' @rdname addLegendSize
+#' @rdname legendSymbols
 addLegendLine <- function(map,
                           pal,
                           values,
@@ -1302,11 +1525,16 @@ addLegendLine <- function(map,
                           fillOpacity = opacity,
                           breaks = 5,
                           baseSize = 10,
-                          numberFormat = function(x) {prettyNum(x, big.mark = ',', scientific = FALSE, digits = 1)},
+                          numberFormat = function(x) {
+                            prettyNum(x, big.mark = ',', scientific = FALSE,
+                                      digits = 1)
+                            },
                           group = NULL,
                           className = 'info legend leaflet-control',
+                          data = leaflet::getMapData(map),
                           ...) {
   shape <- 'rect'
+  values <- parseValues(values = values, data = data)
   sizes <- sizeBreaks(values, breaks, baseSize)
   if ( missing(color) ) {
     stopifnot( missing(color) & !missing(pal))
@@ -1316,7 +1544,7 @@ addLegendLine <- function(map,
     colors <- color
   }
   symbols <- Map(makeSymbol,
-                 shape = 'rect',
+                 shape = shape,
                  width = width,
                  height = sizes,
                  color = 'transparent',
@@ -1324,11 +1552,78 @@ addLegendLine <- function(map,
                  opacity = opacity,
                  fillOpacity = fillOpacity,
                  `stroke-width` = 0)
-  addLegendImage(map, images = symbols, labels = numberFormat(as.numeric(names(sizes))),
+  addLegendImage(map, images = symbols,
+                 labels = numberFormat(as.numeric(names(sizes))),
                  title = title, labelStyle = labelStyle,
                  orientation = orientation, width = width, height = sizes,
                  group = group, className = className, ...)
 
+}
+
+#' @param height
+#'
+#' in pixels
+#'
+#' @export
+#'
+#' @rdname legendSymbols
+addLegendSymbol <- function(map,
+                            pal,
+                            values,
+                            title = NULL,
+                            labelStyle = '',
+                            shape = c('rect', 'circle', 'triangle', 'plus',
+                                      'cross', 'diamond', 'star', 'stadium',
+                                      'polygon'),
+                            orientation = c('vertical', 'horizontal'),
+                            color,
+                            fillColor = color,
+                            strokeWidth = 1,
+                            opacity = 1,
+                            fillOpacity = opacity,
+                            width = 20,
+                            height = width,
+                            group = NULL,
+                            className = 'info legend leaflet-control',
+                            data = leaflet::getMapData(map),
+                            ...
+) {
+  values <- sort(unique(as.factor(parseValues(values, data))))
+  if ( length(levels(values)) > length(shape) ) {
+    stop('values has more factor levels than shape. Maximum levels is 7')
+  }
+  shape <- shape[values]
+  if ( missing(color) ) {
+    stopifnot( missing(color) & !missing(pal))
+    colors <- pal(values)
+  } else {
+    stopifnot(length(color) == 1 || length(color) == length(values))
+    colors <- color
+  }
+  if ( missing(fillColor) ) {
+    if ( !missing(pal) ) {
+      fillColors <- pal(values)
+    } else {
+      fillColors <- colors
+    }
+  } else {
+    stopifnot(length(fillColor) == 1 || length(fillColor) == length(values))
+    fillColors <- fillColor
+  }
+  symbols <- Map(makeSymbol,
+                 shape = shape,
+                 width = width,
+                 height = height,
+                 color = colors,
+                 fillColor = fillColors,
+                 opacity = opacity,
+                 fillOpacity = fillOpacity,
+                 `stroke-width` = strokeWidth)
+  addLegendImage(map, images = symbols,
+                 labels = as.character(values),
+                 title = title, labelStyle = labelStyle,
+                 orientation = orientation, width = width, height = height,
+                 group = group, className = className, ...)
 }
 
 #' Add a legend with Awesome Icons
@@ -1339,7 +1634,8 @@ addLegendLine <- function(map,
 #'
 #' @param iconSet
 #'
-#' a named list from \link[leaflet]{awesomeIconList}, the names will be the labels in the legend
+#' a named list from \link[leaflet]{awesomeIconList}, the names will be the
+#' labels in the legend
 #'
 #' @param title
 #'
@@ -1386,11 +1682,12 @@ addLegendLine <- function(map,
 #'                                    iconRotate = 30,
 #'   ),
 #'   Ionic = makeAwesomeIcon(icon = "ionic", library = "ion",
-#'                           iconColor = '#fffff', markerColor = 'blue',
+#'                           iconColor = '#ffffff', markerColor = 'blue',
 #'                           spin = TRUE,
 #'                           squareMarker = FALSE),
 #'   Glyphicon = makeAwesomeIcon(icon = "plus-sign", library = "glyphicon",
-#'                               iconColor = 'rgb(192, 255, 0)', markerColor = 'darkpurple',
+#'                               iconColor = 'rgb(192, 255, 0)',
+#'                               markerColor = 'darkpurple',
 #'                               spin = TRUE,
 #'                               squareMarker = FALSE)
 #' )
@@ -1422,38 +1719,60 @@ addLegendAwesomeIcon <- function(map,
                                  className = 'info legend leaflet-control',
                                  ...) {
   stopifnot(inherits(iconSet, 'leaflet_awesome_icon_set'))
+  stopifnot( !is.null(names(iconSet)) &&
+               length(names(iconSet)) == length(iconSet) )
   orientation <- match.arg(orientation)
   if ( orientation == 'vertical' ) {
     wrapElements <- htmltools::tags$div
   } else {
     wrapElements <- htmltools::tags$span
   }
-
+  currentDepNames <- vapply(map$dependencies, getElement, name = 'name',
+                            FUN.VALUE = 'character')
+  iconLibraries <- unique(vapply(iconSet, getElement, name = 'library',
+                                 FUN.VALUE = character(1)))
+  verifyIconLibrary(iconLibraries)
+  iconLibraries <- c(
+    fa = 'fontawesome',
+    ion = 'ionicons',
+    glyphicon = 'bootstrap')[iconLibraries]
+  missingDeps <- setdiff(c('leaflet-awesomemarkers', iconLibraries),
+                         currentDepNames)
+  iconDeps <- c(
+    `leaflet-awesomemarkers` = leafletAwesomeMarkersDependencies(),
+    fontawesome = leafletAmFontAwesomeDependencies(),
+    ion = leafletAmIonIconDependencies(),
+    bootstrap = leafletAmBootstrapDependencies())[missingDeps]
+  map$dependencies <- c(map$dependencies, unname(iconDeps))
   htmlElements <-
     Map(icon = iconSet,
         label = names(iconSet),
         f = function(icon, label) {
           markerClass <- ''
           if ( marker ) {
-            markerClass <- sprintf('awesome-marker-icon-%s awesome-marker %s',
-                                   icon[['markerColor']],
-                                   ifelse(icon[['squareMarker']], 'awesome-marker-square', ''))
+            markerClass <- sprintf(
+              'awesome-marker-icon-%s awesome-marker %s',
+              icon[['markerColor']],
+              ifelse(icon[['squareMarker']], 'awesome-marker-square', ''))
           }
       htmltools::tagList(
         wrapElements(
-        htmltools::tags$div(style = 'vertical-align: middle; display: inline-block; position: relative;',
+        htmltools::tags$div(
+          style = 'vertical-align: middle; display: inline-block; position: relative;',
                             class = markerClass,
                             htmltools::tags$i(class = sprintf('%1$s %1$s-%2$s %3$s',
                                                               icon[['library']],
                                                               icon[['icon']],
                                                               ifelse(icon[['spin']], 'fa-spin', '')),
-                                              style = sprintf('color: %s; %s', icon[['iconColor']],
+                                              style = sprintf('color: %s; %s; margin-right: 0px', icon[['iconColor']],
                                                               ifelse(icon[['iconRotate']] == 0, '',
                                                                      sprintf('-webkit-transform: rotate(%1$sdeg);-moz-transform: rotate(%1$sdeg);-o-transform: rotate(%1$sdeg);-ms-transform: rotate(%1$sdeg);transform: rotate(%1$sdeg);',
                                                                              icon[['iconRotate']]))
-                                              )
-                            )#,
-                            #htmltools::tags$span(label, style = sprintf("vertical-align: middle; padding: 1px; %s", 'labelStyle'))
+                                              ),
+                                              if ( !is.null(icon[['text']]) ) {
+                                                icon[['text']]
+                                                }
+                            )
         ),
         htmltools::tags$span(label, style = sprintf('%s', labelStyle))
       )
@@ -1461,9 +1780,11 @@ addLegendAwesomeIcon <- function(map,
     })
   if (!is.null(title)) {
     htmlElements <-
-      append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
+      append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))),
+             after = 0)
   }
-  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements),
+                       className = className, group = group, ...)
 }
 
 leaflegendAddControl <- function(map,
@@ -1473,7 +1794,8 @@ leaflegendAddControl <- function(map,
                                  ...) {
 
   if ( !is.null(group) ) {
-    leafLegendClassName = paste('leaflegend-group', gsub('\\W', '', group), sep = '-')
+    leafLegendClassName <- paste('leaflegend-group', gsub('\\W', '', group),
+                                 sep = '-')
     className <- paste(className, leafLegendClassName)
 
     lf <- leaflet::addControl(map, html = html, className = className, ...)
@@ -1499,5 +1821,46 @@ leaflegendAddControl <- function(map,
                         ")
   } else {
     leaflet::addControl(map, html = html, className = className, ...)
+  }
+}
+
+parseValues <- function(values, data) {
+  if ( inherits(values, 'formula') ) {
+    stopifnot(!is.null(data))
+    leaflet::evalFormula(values, data)
+  } else {
+    values
+  }
+}
+
+# Borrowed from "leaflet" package internal functions
+leafletAmBootstrapDependencies <- function() {
+  list(htmltools::htmlDependency(
+    "bootstrap", "3.3.7", "htmlwidgets/plugins/Leaflet.awesome-markers",
+    package = "leaflet", script = c("bootstrap.min.js"),
+    stylesheet = c("bootstrap.min.css")))
+}
+leafletAmFontAwesomeDependencies <- function() {
+  list(htmltools::htmlDependency(
+    "fontawesome", "4.7.0", "htmlwidgets/plugins/Leaflet.awesome-markers",
+    package = "leaflet", stylesheet = c("font-awesome.min.css")))
+}
+leafletAmIonIconDependencies <- function() {
+  list(htmltools::htmlDependency(
+    "ionicons", "2.0.1", "htmlwidgets/plugins/Leaflet.awesome-markers",
+    package = "leaflet", stylesheet = c("ionicons.min.css")))
+}
+leafletAwesomeMarkersDependencies <- function() {
+  list(htmltools::htmlDependency(
+    "leaflet-awesomemarkers",
+    "2.0.3", "htmlwidgets/plugins/Leaflet.awesome-markers",
+    package = "leaflet", script = c("leaflet.awesome-markers.min.js"),
+    stylesheet = c("leaflet.awesome-markers.css")))
+
+}
+verifyIconLibrary <- function(library) {
+  bad <- library[!(library %in% c("glyphicon", "fa", "ion"))]
+  if (length(bad) > 0) {
+    stop("Invalid icon library names: ", paste(unique(bad), collapse = ", "))
   }
 }